Refactor in domains
