+++ internal/core/domain/partner.go
@@ -1,0 +1,13 @@
+package domain
+
+type Partner struct {
+	Participation float64 `validate:"required"`
+	User          User
+}
+
+func NewPartner(participation float64, user User) *Partner {
+	return &Partner{
+		Participation: participation,
+		User:          user,
+	}
+}
